+++ pac4j-jwt/src/main/java/org/pac4j/jwt/config/signature/RSASignatureConfiguration.java
@@ -2,10 +2,12 @@ package org.pac4j.jwt.config.signature;
 
 import com.nimbusds.jose.JOSEException;
 import com.nimbusds.jose.JWSAlgorithm;
+import com.nimbusds.jose.JWSHeader;
 import com.nimbusds.jose.JWSSigner;
 import com.nimbusds.jose.JWSVerifier;
 import com.nimbusds.jose.crypto.RSASSASigner;
 import com.nimbusds.jose.crypto.RSASSAVerifier;
+import com.nimbusds.jwt.JWTClaimsSet;
 import com.nimbusds.jwt.SignedJWT;
 import lombok.Getter;
 import lombok.Setter;
@@ -78,9 +80,18 @@ public class RSASignatureConfiguration extends AbstractSignatureConfiguration {
 
     /** {@inheritDoc} */
     @Override
-    protected JWSSigner buildSigner() {
+    public SignedJWT sign(JWTClaimsSet claims) {
+        init();
         CommonHelper.assertNotNull("privateKey", privateKey);
-        return new RSASSASigner(this.privateKey);
+
+        try {
+            JWSSigner signer = new RSASSASigner(this.privateKey);
+            val signedJWT = new SignedJWT(new JWSHeader(algorithm), claims);
+            signedJWT.sign(signer);
+            return signedJWT;
+        } catch (final JOSEException e) {
+            throw new TechnicalException(e);
+        }
     }
 
     /** {@inheritDoc} */